+++ channel/handlers_server.go
@@ -3,18 +3,15 @@ package channel
 import (
 	"fmt"
 	"log"
-	"sort"
 	"strconv"
 	"time"
 
 	"github.com/Hucaru/Valhalla/common"
 	"github.com/Hucaru/Valhalla/common/opcode"
 	"github.com/Hucaru/Valhalla/constant"
-	"github.com/Hucaru/Valhalla/constant/skill"
 	"github.com/Hucaru/Valhalla/internal"
 	"github.com/Hucaru/Valhalla/mnet"
 	"github.com/Hucaru/Valhalla/mpacket"
-	"github.com/Hucaru/Valhalla/nx"
 )
 
 func (server *Server) HandleServerPacket(conn mnet.Server, reader mpacket.Reader) {
@@ -554,469 +551,6 @@ func (server *Server) handleMessengerEvent(conn mnet.Server, reader mpacket.Read
 	}
 }
 
-func getAffectedPartyMembers(p *party, src *Player, affected byte) []*Player {
-	if p == nil || src == nil {
-		return nil
-	}
-
-	var total byte
-	for i := 0; i < constant.MaxPartySize; i++ {
-		if p.players[i] != nil {
-			total++
-		}
-	}
-
-	ret := make([]*Player, 0, constant.MaxPartySize)
-
-	for i := 0; i < constant.MaxPartySize; i++ {
-		idx := i + 1
-		mask := partyMemberMaskForIndex(idx, total)
-		if (affected & mask) == 0 {
-			continue
-		}
-
-		member := p.players[i]
-		if member == nil {
-			continue
-		}
-
-		// Must be same map and same instance
-		if member.mapID != src.mapID {
-			continue
-		}
-		if member.inst == nil || src.inst == nil || member.inst.id != src.inst.id {
-			continue
-		}
-
-		// Exclude self
-		if member.ID == src.ID {
-			continue
-		}
-
-		ret = append(ret, member)
-	}
-
-	return ret
-}
-
-func (server *Server) playerSpecialSkill(conn mnet.Client, reader mpacket.Reader) {
-	// Minimal, safe implementation to keep packet stream in sync and apply basic validations/costs.
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	// Dead players cannot cast
-	if plr.hp == 0 {
-		plr.Send(packetPlayerNoChange())
-		return
-	}
-
-	// The packet layout for special skills generally starts with [skillID int32][skillLevel byte]
-	skillID := reader.ReadInt32()
-	skillLevel := reader.ReadByte()
-
-	// Validate the Player owns the skill and level does not exceed learned level
-	ps, ok := plr.skills[skillID]
-	if !ok || skillLevel == 0 || skillLevel > ps.Level {
-		// Possible hack/desync; drop request
-		plr.Send(packetPlayerNoChange())
-		return
-	}
-
-	partyMask := reader.ReadByte() // party flags
-	delay := reader.ReadInt16()    // delay
-
-	switch skill.Skill(skillID) {
-	case skill.Haste, skill.BanditHaste, skill.Bless, skill.IronWill, skill.Rage,
-		skill.Meditation, skill.ILMeditation, skill.MesoUp, skill.HolySymbol, skill.HyperBody, skill.NimbleBody:
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		// Apply to eligible party members in same map/instance per mask
-		if plr.party != nil {
-			affected := getAffectedPartyMembers(plr.party, plr, partyMask)
-			for _, member := range affected {
-				if member == nil {
-					continue
-				}
-
-				member.addForeignBuff(member.ID, skillID, skillLevel, delay)
-				member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-				member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-			}
-		}
-	case skill.SuperGMHaste, skill.SuperGMBless, skill.SuperGMHolySymbol:
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		// Apply buff to the entire map
-		for _, member := range plr.inst.players {
-			if member.ID == plr.ID {
-				continue
-			}
-			member.addForeignBuff(member.ID, skillID, skillLevel, delay)
-			member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-			member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-		}
-	case skill.SuperGMResurrection:
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		for _, member := range plr.inst.players {
-			if member.ID == plr.ID {
-				continue
-			}
-			if member.hp == 0 {
-				member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-				member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-				member.setHP(member.maxHP)
-			}
-		}
-
-	case skill.SuperGMHealDispell:
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		for _, member := range plr.inst.players {
-			if member.ID == plr.ID {
-				continue
-			}
-			member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-			member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-			member.setHP(member.maxHP)
-			member.setMP(member.maxMP)
-			member.buffs.cureAllDebuffs()
-		}
-
-	case skill.Dispel:
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		if plr.party != nil {
-			affected := getAffectedPartyMembers(plr.party, plr, partyMask)
-			for _, member := range affected {
-				if member == nil {
-					continue
-				}
-
-				member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-				member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-				member.buffs.cureAllDebuffs()
-			}
-		}
-
-	case skill.Resurrection:
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		if plr.party != nil {
-			affected := getAffectedPartyMembers(plr.party, plr, partyMask)
-			for _, member := range affected {
-				if member == nil {
-					continue
-				}
-
-				if member.hp == 0 {
-					member.Send(packetPlayerEffectSkill(true, skillID, skillLevel))
-					member.inst.send(packetPlayerSkillAnimation(member.ID, true, skillID, skillLevel))
-					member.setHP(member.maxHP)
-				}
-			}
-		}
-
-	// Nimble feet and recovery beginner skills with cooldown
-	case skill.NimbleFeet, skill.Recovery:
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, true, skillID, skillLevel))
-		// Send cooldown packet for beginner skills
-		if skillData, ok := plr.skills[skillID]; ok {
-			plr.Send(packetPlayerSkillCooldown(skillID, skillData.CooldownTime))
-			// Start timer to clear the cooldown when it expires
-			go func(sid int32, cooldownTime int16, inst *fieldInstance) {
-				time.Sleep(time.Duration(cooldownTime) * time.Second)
-				// Use dispatch pattern for thread safety
-				if inst != nil && inst.dispatch != nil && plr != nil {
-					inst.dispatch <- func() {
-						plr.Send(packetPlayerSkillCooldown(sid, 0))
-					}
-				}
-			}(skillID, skillData.CooldownTime, plr.inst)
-		}
-
-	case skill.DarkSight,
-		skill.MagicGuard,
-		skill.Invincible,
-		skill.SoulArrow, skill.CBSoulArrow,
-		skill.ShadowPartner,
-		skill.MesoGuard,
-		skill.SwordBooster, skill.AxeBooster, skill.PageSwordBooster, skill.BwBooster,
-		skill.SpearBooster, skill.PolearmBooster,
-		skill.BowBooster, skill.CrossbowBooster,
-		skill.ClawBooster, skill.DaggerBooster,
-		skill.SpellBooster, skill.ILSpellBooster,
-		// GM Hide (mapped to invincible bit)
-		skill.SuperGMHide:
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, true, skillID, skillLevel))
-
-	case skill.Threaten,
-		skill.Slow, skill.ILSlow,
-		skill.MagicCrash,
-		skill.PowerCrash,
-		skill.ArmorCrash,
-		skill.ILSeal, skill.Seal,
-		skill.ShadowWeb,
-		skill.Doom:
-
-		_ = reader.ReadInt16() // padding
-
-		mobIDs := make([]int32, 0)
-		restBytes := reader.GetRestAsBytes()
-		for i := 0; i+4 <= len(restBytes); i += 4 {
-			mobID := int32(restBytes[i]) | int32(restBytes[i+1])<<8 | int32(restBytes[i+2])<<16 | int32(restBytes[i+3])<<24
-			if mobID == 0 || mobID < 0 {
-				break
-			}
-			mobIDs = append(mobIDs, mobID)
-		}
-
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-		var statMask int32
-		switch skill.Skill(skillID) {
-		case skill.Threaten:
-			statMask = skill.MobStat.PhysicalDefense | skill.MobStat.MagicDefense
-		case skill.ArmorCrash:
-			statMask = skill.MobStat.PhysicalDefense
-		case skill.PowerCrash:
-			statMask = skill.MobStat.PhysicalDamage
-		case skill.MagicCrash:
-			statMask = skill.MobStat.MagicDefense
-		case skill.Slow, skill.ILSlow:
-			statMask = skill.MobStat.Speed
-		case skill.Seal, skill.ILSeal:
-			statMask = skill.MobStat.Seal
-		case skill.ShadowWeb:
-			statMask = skill.MobStat.Web
-		case skill.Doom:
-			statMask = skill.MobStat.Doom
-		}
-
-		if plr.inst != nil {
-			for _, mobID := range mobIDs {
-				plr.inst.lifePool.applyMobBuff(mobID, skillID, skillLevel, statMask, plr.inst)
-			}
-		}
-
-	case skill.MysticDoor:
-		createMysticDoor(plr, skillID, skillLevel)
-
-	// Summons and puppet:
-	case skill.SummonDragon,
-		skill.SilverHawk, skill.GoldenEagle,
-		skill.Puppet, skill.SniperPuppet:
-		isPuppet := (skill.Skill(skillID) == skill.Puppet || skill.Skill(skillID) == skill.SniperPuppet)
-
-		spawn := plr.pos
-
-		if isPuppet {
-			desiredX := plr.pos.x
-			if (plr.stance & 0x01) == 0 {
-				desiredX += 200 // facing right
-			} else {
-				desiredX -= 200 // facing left
-			}
-			if fld, ok := server.fields[plr.mapID]; ok {
-				if inst, err := fld.getInstance(plr.inst.id); err == nil {
-					snapped := inst.fhHist.getFinalPosition(newPos(desiredX, plr.pos.y, 0))
-					spawn = snapped
-				}
-			}
-		}
-
-		summ := &summon{
-			OwnerID:    plr.ID,
-			SkillID:    skillID,
-			Level:      skillLevel,
-			Pos:        spawn,
-			Stance:     0,
-			Foothold:   spawn.foothold,
-			IsPuppet:   isPuppet,
-			SummonType: 0,
-		}
-
-		if isPuppet {
-			if data, err := nx.GetPlayerSkill(skillID); err == nil {
-				idx := int(skillLevel) - 1
-				if idx >= 0 && idx < len(data) {
-					summ.HP = int(data[idx].X)
-				}
-			}
-		}
-
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.addSummon(summ)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-
-	default:
-		// Always Send a self animation so client shows casting even for non-buffs.
-		plr.addBuff(skillID, skillLevel, delay)
-		plr.inst.send(packetPlayerSkillAnimation(plr.ID, false, skillID, skillLevel))
-	}
-
-	// Apply MP cost/cooldown, if any (reuses the same flow as attack skills).
-	plr.useSkill(skillID, skillLevel, 0)
-	plr.Send(packetPlayerNoChange()) // catch all for things like GM teleport
-}
-
-func (server *Server) playerStopSkill(conn mnet.Client, reader mpacket.Reader) {
-	skillID := reader.ReadInt32()
-
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil || plr.buffs == nil {
-		return
-	}
-
-	cb := plr.buffs
-
-	cb.expireBuffNow(skillID)
-	cb.plr.inst = plr.inst
-	cb.AuditAndExpireStaleBuffs()
-}
-
-func (server *Server) playerCancelBuff(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil || plr.buffs == nil {
-		return
-	}
-
-	cb := plr.buffs
-
-	const grace = 1500 * time.Millisecond
-	now := time.Now().Add(grace).UnixMilli()
-
-	// Collect all timed sources that should be expired by server clock
-	toExpire := make([]int32, 0, len(cb.expireAt))
-	for src, ts := range cb.expireAt {
-		if ts > 0 && ts <= now {
-			toExpire = append(toExpire, src)
-		}
-	}
-
-	for _, src := range toExpire {
-		cb.expireBuffNow(src)
-	}
-
-	// Final sweep for any edge cases
-	cb.plr.inst = plr.inst
-	cb.AuditAndExpireStaleBuffs()
-}
-
-func (server Server) playerSummonMove(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	summonID := reader.ReadInt32()
-	summ := plr.getSummon(summonID)
-	if summ == nil || summ.IsPuppet {
-		return
-	}
-
-	moveData, finalData := parseMovement(reader)
-	moveBytes := generateMovementBytes(moveData)
-
-	summ.Pos = pos{x: finalData.x, y: finalData.y}
-	summ.Stance = finalData.stance
-	summ.Foothold = finalData.foothold
-
-	field, ok := server.fields[plr.mapID]
-	if !ok {
-		return
-	}
-	inst, err := field.getInstance(plr.inst.id)
-	if err != nil {
-		return
-	}
-
-	inst.sendExcept(packetSummonMove(plr.ID, summonID, moveBytes), conn)
-}
-
-func (server *Server) playerSummonDamage(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	summonID := reader.ReadInt32()
-	summ := plr.getSummon(summonID)
-	if summ == nil || !summ.IsPuppet {
-		return
-	}
-
-	_ = int8(reader.ReadByte())
-	damage := reader.ReadInt32()
-	mobID := reader.ReadInt32()
-	_ = reader.ReadByte()
-
-	field, ok := server.fields[plr.mapID]
-	if ok {
-		if inst, e := field.getInstance(plr.inst.id); e == nil {
-			inst.send(packetSummonDamage(plr.ID, summonID, damage, mobID))
-		}
-	}
-
-	if summ.HP-int(damage) < 0 {
-		plr.removeSummon(true, 0x02)
-	} else {
-		summ.HP -= int(damage)
-	}
-}
-
-func (server *Server) playerSummonAttack(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	data, valid := getAttackInfo(reader, *plr, attackSummon)
-	if !valid || len(data.attackInfo) == 0 {
-		return
-	}
-
-	field, ok := server.fields[plr.mapID]
-	if !ok {
-		return
-	}
-	inst, err := field.getInstance(plr.inst.id)
-	if err != nil {
-		return
-	}
-
-	mobDamages := make(map[int32][]int32, len(data.attackInfo))
-	for _, at := range data.attackInfo {
-		if at.spawnID <= 0 || len(at.damages) == 0 {
-			continue
-		}
-		mobDamages[at.spawnID] = append(mobDamages[at.spawnID], at.damages...)
-	}
-	if len(mobDamages) == 0 {
-		return
-	}
-
-	anim := data.action
-	if anim == 0 && len(data.attackInfo) > 0 {
-		anim = data.attackInfo[0].frameIndex
-	}
-
-	inst.sendExcept(packetSummonAttack(plr.ID, data.summonType, anim, byte(len(mobDamages)), mobDamages), conn)
-	for spawnID, damages := range mobDamages {
-		for _, d := range damages {
-			if d > 0 {
-				inst.lifePool.mobDamaged(spawnID, plr, d)
-			}
-		}
-	}
-}
-
 func (server Server) handleGuildEvent(conn mnet.Server, reader mpacket.Reader) {
 	op := reader.ReadByte()
 
@@ -1170,575 +704,6 @@ func (server *Server) guildInviteResult(conn mnet.Server, reader mpacket.Reader)
 		log.Println("Unknown guild invite operation", op, reader)
 	}
 
-}
-
-func (server *Server) playerQuestOperation(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	act := reader.ReadByte()
-	questID := reader.ReadInt16()
-
-	switch act {
-	case constant.QuestStarted:
-		if !plr.tryStartQuest(questID) {
-			plr.Send(packetPlayerNoChange())
-		}
-	case constant.QuestCompleted:
-		if !plr.tryCompleteQuest(questID) {
-			plr.Send(packetPlayerNoChange())
-		}
-	case constant.QuestForfeit:
-		plr.quests.remove(questID)
-		deleteQuest(plr.ID, questID)
-		clearQuestMobKills(plr.ID, questID)
-		plr.Send(packetQuestRemove(questID))
-	case constant.QuestLostItem:
-		count := reader.ReadInt16()
-		questItem := reader.ReadInt16()
-		if count > 0 {
-			if it, err := CreateItemFromID(int32(questItem), count); err == nil {
-				_, _ = plr.GiveItem(it)
-			} else {
-				log.Printf("[QuestPkt] lostItem give failed: err=%v", err)
-			}
-		} else if count < 0 {
-			if !plr.removeItemsByID(int32(questItem), int32(-count)) {
-				log.Printf("[QuestPkt] lostItem remove failed: Item=%d need=%d", questItem, -count)
-			}
-		}
-
-	default:
-		log.Println("Unknown quest operation type:", act)
-	}
-}
-
-func (server *Server) playerFame(conn mnet.Client, reader mpacket.Reader) {
-	source, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	targetID := reader.ReadInt32()
-	up := reader.ReadBool()
-
-	if targetID == source.ID {
-		return
-	}
-
-	target, err := server.players.GetFromID(targetID)
-	if err != nil || target == nil || target.mapID != source.mapID {
-		source.Send(packetFameError(constant.FameIncorrectUser))
-		return
-	}
-
-	if source.level < 15 {
-		source.Send(packetFameError(constant.FameUnderLevel))
-		return
-	}
-
-	if fameHasRecentActivity(source.ID, 24*time.Hour) {
-		source.Send(packetFameError(constant.FameThisDay))
-		return
-	}
-
-	if fameHasRecentActivity(source.ID, 30*24*time.Hour) {
-		source.Send(packetFameError(constant.FameThisMonth))
-		return
-	}
-
-	delta := int16(1)
-	if !up {
-		delta = -1
-	}
-	target.setFame(target.fame + delta)
-
-	if err := fameInsertLog(source.ID, target.ID); err != nil {
-		log.Println("fameInsertLog:", err)
-	}
-
-	target.Send(packetFameNotifyVictim(source.Name, up))
-	source.Send(packetFameNotifySource(target.Name, up, target.fame))
-}
-
-func (server *Server) playerHitReactor(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	spawnID := reader.ReadInt32()
-	_ = reader.ReadInt32() // stance
-	_ = reader.ReadInt16() // delay
-
-	plr.inst.reactorPool.triggerHit(spawnID, 0, server, plr)
-
-}
-
-func (server *Server) playerUseStorage(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	const (
-		actionWithdraw        byte = 4
-		actionDeposit              = 5
-		actionStoreMesos           = 6
-		actionExit                 = 7
-		storageInvFullOrNot        = 9
-		encWithdraw                = 8
-		encDeposit                 = 10
-		storageNotEnoughMesos      = 12
-		storageIsFull              = 13
-		storageDueToAnError        = 14
-		storageSuccess             = 15
-	)
-
-	accountID := conn.GetAccountID()
-	if accountID == 0 {
-		return
-	}
-
-	isRechargeable := func(itemID int32) bool {
-		base := itemID / 10000
-		return base == 207
-	}
-
-	switch reader.ReadByte() {
-	case actionWithdraw:
-		tab := reader.ReadByte()
-		slot := reader.ReadByte()
-
-		stIdx, it := plr.storageInventory.getBySectionSlot(tab, slot)
-		if stIdx < 0 || it == nil || it.ID == 0 {
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		out := *it
-		if !out.isStackable() || out.amount <= 0 {
-			out.amount = 1
-		}
-		out.dbID = 0
-		out.slotID = 0
-
-		if err, _ := plr.GiveItem(out); err != nil {
-			plr.Send(packetNpcStorageResult(storageInvFullOrNot))
-			return
-		}
-
-		plr.storageInventory.removeAt(byte(stIdx))
-		if err := plr.storageInventory.save(plr.accountID); err != nil {
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		sectionItems := plr.storageInventory.getItemsInSection(tab)
-		plr.Send(packetNpcStorageItemsChanged(encWithdraw, plr.storageInventory.maxSlots, tab, 0, sectionItems))
-
-	case actionDeposit:
-		srcSlot := reader.ReadInt16()
-		itemID := reader.ReadInt32()
-		amt := reader.ReadInt16()
-		if amt <= 0 {
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		tab := byte(itemID / 1000000)
-		itemOnChar, getErr := plr.getItem(tab, srcSlot)
-		if getErr != nil || itemOnChar.ID != itemID {
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		if !plr.storageInventory.slotsAvailable() {
-			plr.Send(packetNpcStorageResult(storageIsFull))
-			return
-		}
-
-		if isRechargeable(itemID) {
-			amt = itemOnChar.amount
-		} else if !itemOnChar.isStackable() {
-			amt = 1
-		} else if amt > itemOnChar.amount {
-			amt = itemOnChar.amount
-		}
-
-		storeCopy := itemOnChar
-		storeCopy.amount = amt
-		storeCopy.dbID = 0
-		storeCopy.slotID = 0
-
-		if _, remErr := plr.takeItem(itemID, srcSlot, amt, tab); remErr != nil {
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		if !plr.storageInventory.addItem(storeCopy) {
-			_, _ = plr.GiveItem(storeCopy)
-			plr.Send(packetNpcStorageResult(storageIsFull))
-			return
-		}
-
-		if err := plr.storageInventory.save(plr.accountID); err != nil {
-			_, _ = plr.GiveItem(storeCopy)
-			plr.Send(packetNpcStorageResult(storageDueToAnError))
-			return
-		}
-
-		sectionItems := plr.storageInventory.getItemsInSection(tab)
-		plr.Send(packetNpcStorageItemsChanged(encDeposit, plr.storageInventory.maxSlots, tab, 0, sectionItems))
-
-	case actionStoreMesos:
-		val := reader.ReadInt32()
-		if val < 0 {
-			store := -val
-			if store <= 0 || plr.mesos < store {
-				plr.Send(packetNpcStorageResult(storageNotEnoughMesos))
-				return
-			}
-			plr.giveMesos(-store)
-			if err := plr.storageInventory.changeMesos(store); err != nil {
-				plr.giveMesos(store)
-				plr.Send(packetNpcStorageResult(storageDueToAnError))
-				return
-			}
-			if err := plr.storageInventory.save(plr.accountID); err != nil {
-				_ = plr.storageInventory.changeMesos(-store)
-				plr.giveMesos(store)
-				plr.Send(packetNpcStorageResult(storageDueToAnError))
-				return
-			}
-			plr.Send(packetNpcStorageMesosChanged(storageSuccess, plr.storageInventory.mesos, plr.storageInventory.maxSlots))
-		} else if val > 0 {
-			withdraw := val
-			if err := plr.storageInventory.changeMesos(-withdraw); err != nil {
-				plr.Send(packetNpcStorageResult(storageDueToAnError))
-				return
-			}
-			plr.giveMesos(withdraw)
-			if err := plr.storageInventory.save(plr.accountID); err != nil {
-				_ = plr.storageInventory.changeMesos(withdraw)
-				plr.giveMesos(-withdraw)
-				plr.Send(packetNpcStorageResult(storageDueToAnError))
-				return
-			}
-			plr.Send(packetNpcStorageMesosChanged(storageSuccess, plr.storageInventory.mesos, plr.storageInventory.maxSlots))
-		} else {
-			plr.Send(packetNpcStorageResult(storageIsFull))
-		}
-
-	case actionExit:
-		return
-	}
-}
-
-func (server *Server) playerHandleMessenger(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	mode := reader.ReadByte()
-
-	switch mode {
-	case constant.MessengerEnter:
-		messengerID := reader.ReadInt32()
-		var petAcc int32
-		var vis []internal.KV
-		var hid []internal.KV
-
-		base := make(map[byte]int32)
-		cash := make(map[byte]int32)
-		cashW := int32(0)
-
-		for _, it := range plr.equip {
-			if it.slotID >= 0 {
-				continue
-			}
-
-			slot := byte(-it.slotID)
-			if it.slotID < -100 {
-				slot = byte(-(it.slotID + 100))
-			}
-
-			if slot == 11 {
-				if it.slotID < -100 {
-					cashW = it.ID
-				}
-				continue
-			}
-
-			if it.slotID < -100 {
-				cash[slot] = it.ID
-			} else {
-				if _, ok := base[slot]; !ok {
-					base[slot] = it.ID
-				}
-			}
-		}
-
-		order := func(m map[byte]int32) []byte {
-			ks := make([]byte, 0, len(m))
-			for k := range m {
-				ks = append(ks, k)
-			}
-			sort.Slice(ks, func(i, j int) bool { return ks[i] < ks[j] })
-			return ks
-		}
-
-		for _, k := range order(base) {
-			if v, ok := cash[k]; ok {
-				vis = append(vis, internal.KV{K: k, V: v})
-				hid = append(hid, internal.KV{K: k, V: base[k]})
-			} else {
-				vis = append(vis, internal.KV{K: k, V: base[k]})
-			}
-		}
-
-		for _, k := range order(cash) {
-			if _, used := base[k]; !used {
-				vis = append(vis, internal.KV{K: k, V: cash[k]})
-			}
-		}
-
-		p := internal.PacketMessengerEnter(plr.ID, messengerID, plr.face, plr.hair, cashW, petAcc, server.id, plr.gender, plr.skin, plr.Name, vis, hid)
-		server.world.Send(p)
-	case constant.MessengerLeave:
-		p := internal.PacketMessengerLeave(plr.ID)
-		server.world.Send(p)
-	case constant.MessengerInvite:
-		invitee := reader.ReadString(reader.ReadInt16())
-		p := internal.PacketMessengerInvite(plr.ID, server.id, plr.Name, invitee)
-		server.world.Send(p)
-	case constant.MessengerBlocked:
-		invitee := reader.ReadString(reader.ReadInt16())
-		inviter := reader.ReadString(reader.ReadInt16())
-		blockMode := reader.ReadByte()
-		p := internal.PacketMessengerBlocked(plr.ID, server.id, blockMode, plr.Name, invitee, inviter)
-		server.world.Send(p)
-	case constant.MessengerChat:
-		message := reader.ReadString(reader.ReadInt16())
-		p := internal.PacketMessengerChat(plr.ID, server.id, plr.Name, message)
-		server.world.Send(p)
-	case constant.MessengerAvatar:
-		var petAcc int32
-		var vis []internal.KV
-		var hid []internal.KV
-
-		base := make(map[byte]int32)
-		cash := make(map[byte]int32)
-		cashW := int32(0)
-
-		for _, it := range plr.equip {
-			if it.slotID >= 0 {
-				continue
-			}
-
-			slot := byte(-it.slotID)
-			if it.slotID < -100 {
-				slot = byte(-(it.slotID + 100))
-			}
-
-			if slot == 11 {
-				if it.slotID < -100 {
-					cashW = it.ID
-				}
-				continue
-			}
-
-			if it.slotID < -100 {
-				cash[slot] = it.ID
-			} else {
-				if _, ok := base[slot]; !ok {
-					base[slot] = it.ID
-				}
-			}
-		}
-
-		order := func(m map[byte]int32) []byte {
-			ks := make([]byte, 0, len(m))
-			for k := range m {
-				ks = append(ks, k)
-			}
-			sort.Slice(ks, func(i, j int) bool { return ks[i] < ks[j] })
-			return ks
-		}
-
-		for _, k := range order(base) {
-			if v, ok := cash[k]; ok {
-				vis = append(vis, internal.KV{K: k, V: v})
-				hid = append(hid, internal.KV{K: k, V: base[k]})
-			} else {
-				vis = append(vis, internal.KV{K: k, V: base[k]})
-			}
-		}
-
-		for _, k := range order(cash) {
-			if _, used := base[k]; !used {
-				vis = append(vis, internal.KV{K: k, V: cash[k]})
-			}
-		}
-
-		p := internal.PacketMessengerAvatar(plr.gender, plr.skin, server.id, plr.ID, plr.face, plr.hair, cashW, petAcc, plr.Name, vis, hid)
-		server.world.Send(p)
-	default:
-		return
-	}
-
-}
-
-func (server *Server) playerPetSpawn(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	slot := reader.ReadInt16()
-
-	petItem, err := plr.getItem(5, slot)
-	if !petItem.pet || err != nil {
-		plr.Send(packetPlayerNoChange())
-		return
-	}
-
-	if petItem.petData == nil {
-		sn, _ := nx.GetCommoditySNByItemID(petItem.ID)
-		petItem.petData = newPet(petItem.ID, sn, petItem.dbID)
-		savePet(&petItem)
-	}
-
-	petEquipped := plr.petCashID != 0
-	changePet := petEquipped && plr.petCashID == int64(petItem.petData.sn)
-
-	if petEquipped {
-		plr.inst.send(packetPetRemove(plr.ID, constant.PetRemoveNone))
-		plr.petCashID = 0
-	}
-
-	if !changePet {
-		plr.petCashID = int64(petItem.petData.sn)
-
-		if plr.pet == nil || plr.pet.sn != petItem.petData.sn {
-			plr.pet = petItem.petData
-		}
-
-		plr.pet.pos = plr.pos
-		plr.inst.send(packetPetSpawn(plr.ID, plr.pet))
-
-		if plr.pet.spawnDate == 0 {
-			plr.Send(packetPlayerPetUpdate(plr.pet.sn))
-		}
-		plr.pet.spawnDate = time.Now().Unix()
-		plr.pet.spawned = true
-	}
-
-	plr.MarkDirty(DirtyPet, time.Millisecond*300)
-	plr.Send(packetPlayerNoChange())
-}
-
-func (server *Server) playerPetMove(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	moveData, finalData := parseMovement(reader)
-	moveBytes := generateMovementBytes(moveData)
-
-	plr.pet.updateMovement(finalData)
-
-	field, ok := server.fields[plr.mapID]
-
-	if !ok {
-		return
-	}
-
-	inst, err := field.getInstance(plr.inst.id)
-
-	if err != nil {
-		return
-	}
-
-	inst.movePlayerPet(plr.ID, moveBytes, plr)
-}
-
-func (server *Server) playerPetAction(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	actType := reader.ReadByte()
-	act := reader.ReadByte()
-	msg := reader.ReadRestAsString()
-
-	plr.inst.send(packetPetAction(plr.ID, actType, act, msg))
-}
-
-func (server *Server) playerPetInteraction(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil {
-		return
-	}
-
-	if plr.pet == nil || !plr.pet.spawned {
-		return
-	}
-
-	doMultiplier := reader.ReadByte()
-	interactionID := reader.ReadByte()
-
-	petItem := plr.pet
-	success := handlePetInteraction(plr, petItem, interactionID, doMultiplier == 1)
-	plr.Send(packetPetInteraction(plr.ID, interactionID, success, false))
-}
-
-func (server *Server) playerPetLoot(conn mnet.Client, reader mpacket.Reader) {
-	plr, err := server.players.GetFromConn(conn)
-	if err != nil || plr.pet == nil || !plr.pet.spawned {
-		return
-	}
-
-	reader.Skip(4) // unused pos
-	dropID := reader.ReadInt32()
-
-	err, drop := plr.inst.dropPool.findDropFromID(dropID)
-	if err != nil {
-		// Pet spams pickup so just silently return :)
-		return
-	}
-
-	if plr.pet.pos.x-drop.finalPos.x > 800 || plr.pet.pos.y-drop.finalPos.y > 600 {
-		// Hax
-		log.Printf("Player: %s pet tried to pickup an item from far away", plr.Name)
-		plr.Send(packetDropNotAvailable())
-		plr.Send(packetInventoryDontTake())
-		return
-	}
-
-	if !plr.petCanTakeDrop(drop) {
-		return
-	}
-
-	if drop.mesos > 0 {
-		amount := int32(plr.inst.dropPool.rates.mesos * float32(drop.mesos))
-		plr.giveMesos(amount)
-	} else {
-		err, _ = plr.GiveItem(drop.item)
-		if err != nil {
-			plr.Send(packetInventoryFull())
-			plr.Send(packetInventoryDontTake())
-			return
-		}
-
-	}
-
-	plr.inst.dropPool.playerAttemptPickup(drop, plr, 5)
 }
 
 func (server *Server) handleCharacterDeleted(conn mnet.Server, reader mpacket.Reader) {
